feat: Update routes and navigation components
