+++ src/constants.ts
@@ -1,5 +1,6 @@
 const ROUTER_PATHS = {
-  HOME: '/',
+  ONBORDING: '/',
+  CREATE_JOIN_INFO: '/create-join-info',
   PHOTO_LIST: '/photo-list',
   SETTING: '/setting',
   SETTING_VRCHAT_LOG_PATH: '/setting/vrchat-log-path',
